pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
pytest = "^8.3.3"
python-dotenv = "^1.0.1"
yfinance = "^0.2.43"
yahoo-fin = "^0.8.9.1"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
mypy = "^1.11.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 119 # Максимальная длина строки
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''  # Файлы, которые не нужно форматировать

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'